requirements.txt
alembic
cryptography
fastapi
passlib
PyJWT
PyMySQL
python-dotenv
requests
SQLAlchemy
uvicorn